rename serverAddress, add 'getToken' function
