--- piktorlog/src/store/serverAddress.js
@@ -1,1 +1,0 @@
-export default process.env.NODE_ENV === 'development' ? 'http://localhost:4000' : 'https://piktorlog.herokuapp.com/';
